+++ Hawk/functions/User/Get-HawkUserMailSendActivity.ps1
@@ -1,20 +1,20 @@
-Function Get-HawkUserExchangeSearchQuery {
+Function Get-HawkUserMailSendActivity {
     <#
     .SYNOPSIS
-        This will export SearchQueryInitiatedExchange operations from the Unified Audit Log (UAL). Must be connected to Exchange Online
+        This will export Send operations from the Unified Audit Log (UAL). Must be connected to Exchange Online
         using the Connect-EXO or Connect-ExchangeOnline module. M365 E5 or G5 license is required for this function to work.
         This telemetry will ONLY be availabe if Advanced Auditing is enabled for the M365 user.
     .DESCRIPTION
-        This function queries for searches performed in Exchange, providing visibility into what users are searching for and potential 
-        indications of insider threats or data exploration during incidents.
+        This function queries for message-sending activity within Exchange, providing visibility into outbound communications 
+        that could be relevant for identifying data exfiltration attempts, phishing campaigns, or other malicious activity.
     .PARAMETER UserPrincipalName
         Specific user(s) to be investigated
     .EXAMPLE
-        Get-HawkUserExchangeSearchQuery -UserPrincipalName bsmith@contoso.com
-        Returns search queries from Unified Audit Log (UAL) that correspond to the UserPrincipalName that is provided
+        Get-HawkUserMailSendActivity -UserPrincipalName bsmith@contoso.com
+        Returns send activity queries from Unified Audit Log (UAL) that correspond to the UserPrincipalName that is provided
     .OUTPUTS
-        ExchangeSearchQueries_bsmith@contoso.com.csv /json
-        Simple_ExchangeSearchQueries_bsmith@contoso.com.csv/json
+        SendActivity_bsmith@contoso.com.csv /json
+        Simple_SendActivity_bsmith@contoso.com.csv/json
     
     .LINK
         https://www.microsoft.com/security/blog/2020/12/21/advice-for-incident-responders-on-recovery-from-systemic-identity-compromises/
@@ -35,7 +35,7 @@ Function Get-HawkUserExchangeSearchQuery {
             if (Test-HawkGlobalObject) {
                 Initialize-HawkGlobalObject
             }
-            Out-LogFile "Starting Unified Audit Log (UAL) search for 'SearchQueryInitiatedExchange'" -Action
+            Out-LogFile "Starting Unified Audit Log (UAL) search for mail'Send' logs" -Action
             Out-LogFile "Please be patient, this can take a while..." -Information
             Test-EXOConnection
         }#End Begin
@@ -49,14 +49,14 @@ Function Get-HawkUserExchangeSearchQuery {
                 [string]$User = $UserObject.UserPrincipalName
                 
                 # Verify that user has operation enabled for auditing. Otherwise, move onto next user.
-                if (Test-OperationEnabled -User $User -Operation 'SearchQueryInitiated') {
-                    Out-LogFile "Operation 'SearchQueryInitiated' verified enabled for $User." -Information
+                if (Test-OperationEnabled -User $User -Operation 'Send') {
+                    Out-LogFile "Operation 'Send' verified enabled for $User." -Information
                     try {
-                        #Retrieve all audit data for Exchange search queries 
-                        $SearchCommand = "Search-UnifiedAuditLog -Operations 'SearchQueryInitiatedExchange' -UserIds $User"
-                        $ExchangeSearches = Get-AllUnifiedAuditLogEntry -UnifiedSearch $SearchCommand
+                        #Retrieve all audit data for Exchange send activity
+                        $SearchCommand = "Search-UnifiedAuditLog -Operations 'Send' -UserIds $User"
+                        $ExchangeSends = Get-AllUnifiedAuditLogEntry -UnifiedSearch $SearchCommand
                         
-                        if ($ExchangeSearches.Count -gt 0){
+                        if ($ExchangeSends.Count -gt 0){
                             
                             #Define output directory path for user
                             $UserFolder = Join-Path -Path $Hawk.FilePath -ChildPath $User
@@ -67,21 +67,21 @@ Function Get-HawkUserExchangeSearchQuery {
                             }
         
                             #Compress raw data into more simple view
-                            $ExchangeSearchesSimple = $ExchangeSearches | Get-SimpleUnifiedAuditLog
+                            $ExchangeSendsSimple = $ExchangeSends | Get-SimpleUnifiedAuditLog
         
                             #Export both raw and simplistic views to specified user's folder
-                            $ExchangeSearches | Select-Object -ExpandProperty AuditData | Convertfrom-Json | Out-MultipleFileType -FilePrefix "ExchangeSearchQueries_$User" -User $User -csv -json
-                            $ExchangeSearchesSimple | Out-MultipleFileType -FilePrefix "Simple_ExchangeSearchQueries_$User" -User $User -csv -json
+                            $ExchangeSends | Select-Object -ExpandProperty AuditData | Convertfrom-Json | Out-MultipleFileType -FilePrefix "SendActivity_$User" -User $User -csv -json
+                            $ExchangeSendsSimple | Out-MultipleFileType -FilePrefix "Simple_SendActivity_$User" -User $User -csv -json
                         } else {
-                            Out-LogFile "Get-HawkUserExchangeSearchQuery completed successfully" -Information
+                            Out-LogFile "Get-HawkUserMailSendActivity completed successfully" -Information
                             Out-LogFile "No items found for $User." -Information
                         }
                     } catch {
-                        Out-LogFile "Error processing Exchange Search Queries for $User : $_" -isError
+                        Out-LogFile "Error processing Send Activity for $User : $_" -isError
                         Write-Error -ErrorRecord $_ -ErrorAction Continue
                     }
                 } else {
-                    Out-LogFile "Operation 'SearchQueryInitiated' is not enabled for $User." -Information
+                    Out-LogFile "Operation 'Send' is not enabled for $User." -Information
                     Out-LogFile "No data recorded for $User." -Information
                 }
             }
@@ -89,7 +89,7 @@ Function Get-HawkUserExchangeSearchQuery {
         }#End Process
     
         END{
-            Out-Logfile "Completed exporting Search Query logs" -Information
+            Out-Logfile "Completed exporting Send Activity logs" -Information
         }#End End
     
     }