+++ Hawk/changelog.md
@@ -99,3 +99,4 @@
 - Add ability to pass command line arguments to Start-HawkUserInvestigation and Start-HawkTenantInvestigation
 - Added search of Exchange Search Queries to the User Investigation (Get-HawkUserExchangeSearchQuery)
 - Implemented check to verify that an Exchange operation is enabled for auditing before attempting to pull logs
+- Added log pull of user Send activity to the User Investigation (Get-HawkUserMailSendActivity)